chore(testsite): refactor code
